Created a new Xcode project called SAAnimationView Added a new framework target to that project called SAAnimationView Added another target to that project called Example and removed the old AnimationExamples directory. This Example target actually makes use of the framework generated by the SAAnimationView target. Placed all the source files for the framework under a Sources directory.
