--- AnimationExamples/AnimationExamples/AppDelegate.h
@@ -1,17 +1,0 @@
-//
-//  AppDelegate.h
-//  AnimationExamples
-//
-//  Created by Emilio Peláez on 5/20/15.
-//  Copyright (c) 2015 Savvy Apps. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-@interface AppDelegate : UIResponder <UIApplicationDelegate>
-
-@property (strong, nonatomic) UIWindow *window;
-
-
-@end
-
