--- AnimationExamples/AnimationExamples/main.m
@@ -1,16 +1,0 @@
-//
-//  main.m
-//  AnimationExamples
-//
-//  Created by Emilio Peláez on 5/20/15.
-//  Copyright (c) 2015 Savvy Apps. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-#import "AppDelegate.h"
-
-int main(int argc, char * argv[]) {
-	@autoreleasepool {
-	    return UIApplicationMain(argc, argv, nil, NSStringFromClass([AppDelegate class]));
-	}
-}
