--- AnimationExamples/AnimationExamplesTests/AnimationExamplesTests.m
@@ -1,40 +1,0 @@
-//
-//  AnimationExamplesTests.m
-//  AnimationExamplesTests
-//
-//  Created by Emilio Peláez on 5/20/15.
-//  Copyright (c) 2015 Savvy Apps. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-#import <XCTest/XCTest.h>
-
-@interface AnimationExamplesTests : XCTestCase
-
-@end
-
-@implementation AnimationExamplesTests
-
-- (void)setUp {
-    [super setUp];
-    // Put setup code here. This method is called before the invocation of each test method in the class.
-}
-
-- (void)tearDown {
-    // Put teardown code here. This method is called after the invocation of each test method in the class.
-    [super tearDown];
-}
-
-- (void)testExample {
-    // This is an example of a functional test case.
-    XCTAssert(YES, @"Pass");
-}
-
-- (void)testPerformanceExample {
-    // This is an example of a performance test case.
-    [self measureBlock:^{
-        // Put the code you want to measure the time of here.
-    }];
-}
-
-@end
