+++ Example/Animation Examples/CIMenuCloseAnimationView.h
