+++ Example/Animation Examples/CIMenuCloseAnimationView.m
