+++ Example/Animation Examples/Extra Examples/SAColorAnimationView.h
