+++ Example/Animation Examples/Extra Examples/SAColorAnimationView.m
