+++ Example/Animation Examples/SAPlayPauseAnimationView.h
