+++ Example/Animation Examples/SAPlayPauseAnimationView.m
