+++ Example/AppDelegate.h
@@ -1,0 +1,15 @@
+//
+//  AppDelegate.h
+//  Example
+//
+//  Created by Arnold Plakolli on 4.5.21.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AppDelegate : UIResponder <UIApplicationDelegate, UIWindowSceneDelegate>
+
+@property (strong, nonatomic) UIWindow * window;
+  
+@end
+
