+++ Example/AppDelegate.m
@@ -1,0 +1,20 @@
+//
+//  AppDelegate.m
+//  Example
+//
+//  Created by Arnold Plakolli on 4.5.21.
+//
+
+#import "AppDelegate.h"
+
+@interface AppDelegate ()
+
+@end
+
+@implementation AppDelegate
+
+- (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
+  return YES;
+}
+
+@end
