+++ Example/Assets.xcassets/first.imageset/first.pdf
